+++ tutorials/tutorial8/src/Views.js
@@ -252,10 +252,10 @@ export class AvatarPawn extends mix(Pawn).with(PM_Smoothed, PM_ThreeVisible, PM_
             const yawQ = q_axisAngle([0,1,0], this.yaw);
             const t = v3_scale([this.left + this.right, 0, (this.fore + this.back)], 5 * delta/1000);
             const tt = v3_rotate(t, yawQ);
-            let translation = v3_add(this.translation, tt);
+            const translation = v3_add(this.translation, tt);
             this.positionTo(translation, yawQ);
-            this.say("viewGlobalChanged");
-            this.refreshCameraTransform(); // Required by PM_ThreeCamera to trigger a camera refresh when you're driving.
+            // this.say("viewGlobalChanged");
+            // this.refreshCameraTransform(); // Required by PM_ThreeCamera to trigger a camera refresh when you're driving.
         }
     }