+++ packages/kernel/CHANGELOG.md
@@ -34,6 +34,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ViewAssetCache
 - Generic RenderManager
 - Rapier physics remove from full package
+- "viewGlobalChanged" event
+- lookGlobal pawn property
 
 ## [1.4.0] - 2022-10-3
 ### Added